requirements.txt
pyserial>=3.5
matplotlib>=3.8
pandas>=2.2
python-docx>=1.1
